feat: add frontend for bloglist and enhance backend functionality
- Added a new frontend for the bloglist application using React and Vite.
    - Created components such as `App.jsx` and `Blog.jsx`.
    - Configured Vite for development.
    - Added ESLint configuration and `.gitignore` for the frontend.
- Enhanced backend functionality:
    - Added a new `login.js` controller for handling authentication.
    - Updated existing controllers (`blogs.js`, `users.js`) and models (`blog.js`, `user.js`) for improved functionality.
    - Added new REST request files for testing (`login_request.rest`).
    - Updated middleware for better error handling.
- Updated dependencies in `package.json` and `package-lock.json` for both backend and frontend.

+++ back-blogsapp/controllers/blogs.js
@@ -1,19 +1,29 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const mongoose = require('mongoose')
+const { userExtractor } = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find()
+  const blogs = await Blog.find().populate('user', { blogs: 0 })
   response.status(200).json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
-  const newBlog = request.body
-  if (!newBlog.hasOwnProperty('likes')) {
-    newBlog.likes = 0
-  }
-  const blog = new Blog(newBlog)
-  const savedBlog= await blog.save()
+blogsRouter.post('/', userExtractor, async (request, response) => {
+  const { title, author, url, likes } = request.body
+  const user = request.user
+
+  const blog = new Blog({
+    title,
+    author,
+    url,
+    likes: likes || 0,
+    user: user._id
+  })
+
+  const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
   response.status(201).json(savedBlog)
 })
 
@@ -42,18 +52,29 @@ blogsRouter.put('/:id', async (request, response) => {
   response.status(200).json(updatedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
   const id = request.params.id
+  const user = request.user
 
   if (!mongoose.Types.ObjectId.isValid(id)) {
     return response.status(400).json({ error: 'Invalid ID format' })
   }
 
-  const deletedBlog = await Blog.findByIdAndDelete(id)
+  const blog = await Blog.findById(id)
 
-  if (!deletedBlog) {
+  if (!blog) {
     return response.status(404).json({ message: 'resource not found' })
   }
+
+  if (blog.user.toString() !== user._id.toString()) {
+    return response.status(403).json({
+      error: 'Current user doesn\'t have permissions to delete this blog'
+    })
+  }
+
+  await Blog.findByIdAndDelete(id)
+  console.log(user)
+  user.blogs = user.blogs.filter(blogId => blogId.toString() !== id)
 
   response.status(204).end()
 })
